Invert the hierarchy of status and buttons
Wrap the cell buttons in a status element rather than the other way
round so that extra information can be passed to the button in the
future.  We might want to pass down the enabled state of the button
from the server, but we can't make use of that if the status element is
inside the button without some magical hackery.