Add ids to cards
So that the line of the error popups can track them correctly.  Without
an ID per card, the line can point to the wrong card if another card is
played because the cards later in the hand than the played card move
into a status item with a lower ID when the user's hand channel is
updated.